refactor: streamline QR code bulk operation tracking by centralizing event property construction and validation logic
