+++ src/utils/qrCodeBulkOperationTracking.ts
@@ -6,9 +6,9 @@
  *
  * @module src/utils/qrCodeBulkOperationTracking
  */
-import posthogClient, { PostHogEventProperties } from "@/utils/posthogClient";
-import { getBaseEventProperties } from "@/utils/conversionTrackingEventUtils";
+import posthogClient from "@/utils/posthogClient";
 import { QrCodeBulkOperationProperties } from "@/interfaces/conversionTrackings/QrCodeBulkOperationProperties";
+import { buildQrCodeBulkOperationEvent } from "@/utils/qrCodeBulkOperationEvent";
 
 /**
  * Track QR code bulk operation event
@@ -16,33 +16,33 @@ import { QrCodeBulkOperationProperties } from "@/interfaces/conversionTrackings/
  */
 export const trackQrCodeBulkOperation = (
   properties: QrCodeBulkOperationProperties
-): void => {
+): boolean => {
   if (
     !properties.operation_type ||
     !["bulk_download", "bulk_delete"].includes(properties.operation_type)
   ) {
     console.warn("Invalid operation_type provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (!properties.qr_codes_count || properties.qr_codes_count <= 0) {
     console.warn("Invalid qr_codes_count provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (!properties.qr_code_ids || properties.qr_code_ids.length === 0) {
     console.warn("Invalid qr_code_ids provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (!properties.url_ids || properties.url_ids.length === 0) {
     console.warn("Invalid url_ids provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (!properties.qr_code_titles || properties.qr_code_titles.length === 0) {
     console.warn("Invalid qr_code_titles provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (!properties.short_urls || properties.short_urls.length === 0) {
     console.warn("Invalid short_urls provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
   if (
     !properties.customization_options ||
@@ -51,7 +51,7 @@ export const trackQrCodeBulkOperation = (
     console.warn(
       "Invalid customization_options provided to trackQrCodeBulkOperation"
     );
-    return;
+    return false;
   }
   if (
     !["list_view", "grid_view", "bulk_actions_panel"].includes(
@@ -59,7 +59,7 @@ export const trackQrCodeBulkOperation = (
     )
   ) {
     console.warn("Invalid ui_source provided to trackQrCodeBulkOperation");
-    return;
+    return false;
   }
 
   // Validate operation-specific properties
@@ -71,7 +71,7 @@ export const trackQrCodeBulkOperation = (
       console.warn(
         "Invalid download_format provided for bulk_download operation"
       );
-      return;
+      return false;
     }
     if (
       !properties.download_method ||
@@ -80,7 +80,7 @@ export const trackQrCodeBulkOperation = (
       console.warn(
         "Invalid download_method provided for bulk_download operation"
       );
-      return;
+      return false;
     }
   }
 
@@ -92,31 +92,11 @@ export const trackQrCodeBulkOperation = (
       console.warn(
         "Invalid deletion_method provided for bulk_delete operation"
       );
-      return;
+      return false;
     }
   }
 
-  const sanitizedProperties = {
-    ...properties,
-    qr_code_ids: JSON.stringify(properties.qr_code_ids),
-    url_ids: JSON.stringify(properties.url_ids),
-    qr_code_titles: JSON.stringify(
-      properties.qr_code_titles.map((title) => title?.substring(0, 100))
-    ),
-    short_urls: JSON.stringify(
-      properties.short_urls.map((url) => url?.substring(0, 200))
-    ),
-    customization_options: JSON.stringify(
-      properties.customization_options.map((options) => JSON.stringify(options))
-    ),
-    deletion_reason: properties.deletion_reason?.substring(0, 200),
-    error_message: properties.error_message?.substring(0, 200),
-  };
-
-  const eventProperties: PostHogEventProperties = {
-    ...getBaseEventProperties(),
-    ...sanitizedProperties,
-  };
-
-  posthogClient.captureEvent("qr_code_bulk_operation", eventProperties);
+  const built = buildQrCodeBulkOperationEvent(properties);
+  posthogClient.captureEvent("qr_code_bulk_operation", built.event);
+  return true;
 };
